moving shuffle functions to common place
