+++ test/ck_tile/gemm_block_scale/test_gemm_quant_base.hpp
@@ -133,59 +133,6 @@ class TestCkTileGemmQuantBase : public ::testing::Test
         // Use higher threshold
         return ck_tile::make_tuple(std::max(rtol, rtol_split_k), std::max(atol, atol_split_k));
     }
-
-    template <typename T>
-    auto shuffle_b(const ck_tile::HostTensor<T>& t)
-    {
-        assert(t.get_lengths().size() == 2);
-        int n_                = t.get_lengths()[1];
-        int k_                = t.get_lengths()[0];
-        constexpr int divisor = N_Warp_Tile == 32 ? 2 : 4;
-        ck_tile::HostTensor<T> t_view(
-            {n_ / N_Warp_Tile, N_Warp_Tile, k_ / K_Warp_Tile, divisor, K_Warp_Tile / divisor});
-        std::copy(t.begin(), t.end(), t_view.begin());
-        return ck_tile::reference_permute(t_view, {0, 2, 3, 1, 4});
-    }
-
-    template <typename T>
-    auto shuffle_bq_permuteN(const ck_tile::HostTensor<T>& t)
-    {
-    assert(t.get_lengths().size() == 2);
-
-        int n_                  = t.get_lengths()[1];
-        int bqk_                = t.get_lengths()[0];
-        constexpr int NRepeat = N_Tile / N_Warp_Tile / N_Warp;
-
-        ck_tile::HostTensor<T> t_view({n_ / N_Tile,
-                                    N_Warp,
-                                    N_Warp_Tile,
-                                    NRepeat,
-                                    bqk_});
-        std::copy(t.begin(), t.end(), t_view.begin());
-        return ck_tile::reference_permute(t_view, {0, 3, 1, 2, 4});
-    }
-
-    template <typename T>
-    auto shuffle_b_permuteN(const ck_tile::HostTensor<T>& t)
-    {
-        assert(t.get_lengths().size() == 2);
-
-        int n_                = t.get_lengths()[1];
-        int k_                = t.get_lengths()[0];
-        constexpr int divisor = N_Warp_Tile == 32 ? 2 : 4;
-        constexpr int NRepeat = N_Tile / N_Warp_Tile / N_Warp;
-
-        ck_tile::HostTensor<T> t_view({n_ / N_Tile,
-                                    N_Warp,
-                                    N_Warp_Tile,
-                                    NRepeat,
-                                    k_ / K_Warp_Tile,
-                                    divisor,
-                                    K_Warp_Tile / divisor});
-
-        std::copy(t.begin(), t.end(), t_view.begin());
-        return ck_tile::reference_permute(t_view, {0, 3, 1, 4, 5, 2, 6});
-    }
 };
 
 // Define generic QuantTypeTraits template (will be specialized)
